+++ opensquirrel/passes/mapper/general_mapper.py
@@ -6,7 +6,7 @@ from abc import ABC, abstractmethod
 from typing import TYPE_CHECKING, Any
 
 if TYPE_CHECKING:
-    from opensquirrel.ir import IR
+    from opensquirrel import Circuit
     from opensquirrel.passes.mapper.mapping import Mapping
 
 
@@ -16,4 +16,4 @@ class Mapper(ABC):
     def __init__(self, **kwargs: Any) -> None: ...
 
     @abstractmethod
-    def map(self, ir: IR, qubit_register_size: int) -> Mapping: ...
+    def map(self, circuit: Circuit, qubit_register_size: int) -> Mapping: ...