+++ opensquirrel/passes/mapper/simple_mappers.py
@@ -14,7 +14,7 @@ from opensquirrel.passes.mapper.general_mapper import Mapper
 from opensquirrel.passes.mapper.mapping import Mapping
 
 if TYPE_CHECKING:
-    from opensquirrel.ir import IR
+    from opensquirrel import Circuit
 
 
 class IdentityMapper(Mapper):
@@ -22,7 +22,11 @@ class IdentityMapper(Mapper):
         """An ``IdentityMapper`` maps each virtual qubit to exactly the same physical qubit."""
         super().__init__(**kwargs)
 
-    def map(self, ir: IR, qubit_register_size: int) -> Mapping:
+    def map(
+        self,
+        circuit: Circuit,
+        qubit_register_size: int,
+    ) -> Mapping:
         """Map the circuit according to a identity mapping.
 
         Args:
@@ -48,16 +52,7 @@ class HardcodedMapper(Mapper):
         self._mapping = mapping
 
     def map(self, ir: IR, qubit_register_size: int) -> Mapping:
-        """Map the circuit according to the hardcoded mapping.
-
-        Args:
-            ir (IR): The intermediate representation of the quantum circuit to be mapped.
-            qubit_register_size (int): The size of the (virtual) qubit register.
-
-        Returns:
-            Mapping from virtual to physical qubits.
-
-        """
+        """Return the hardcoded mapping."""
         if qubit_register_size != self._mapping.size():
             msg = (
                 f"the size of the mapping {self._mapping.size()!r} is not equal to the number of qubits"
@@ -79,16 +74,7 @@ class RandomMapper(Mapper):
         self.seed = seed
 
     def map(self, ir: IR, qubit_register_size: int) -> Mapping:
-        """Map the circuit according to the random mapping.
-
-        Args:
-            ir (IR): The intermediate representation of the quantum circuit to be mapped.
-            qubit_register_size (int): The size of the (virtual) qubit register.
-
-        Returns:
-            Mapping from virtual to physical qubits.
-
-        """
+        """Create a random mapping."""
         if self.seed:
             random.seed(self.seed)